Standardize us of pass-by-value vs. pass-by-pointer
Handler-associated functions mostly injest arguments by value and return by pointer. This mostly prevents side effects stemming from mutation, but also allows for `nil` to be returned (in place of the expected result) when an error occurs.

If an argument to a function is a pointer after this commit, it is either because it is useful to check whether that parameter is nil, or because it is intentionally mutable (in this function or downstream of it.) It may alternately be because a downstream function (ex. from an external library such as pop) requires a pointer rather than value argument.